% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bjpr.R
\name{bjpr}
\alias{bjpr}
\title{Create a PRC for each model}
\usage{
bjpr(testy, x1, x2)
}
\arguments{
\item{testy}{Observed outcome values of test dataset}

\item{x1}{predicted values of a model}

\item{x2}{model name}
}
\value{
PRC plot
}
\description{
Create a PRC for each model
}
\examples{
testy<-sample(c(0,1), 154, replace=TRUE)
pred_test1<-runif (154, min=0, max=1)
pred_test2<- runif (154, min=0, max=1)
apr=bjpr(testy, pred_test1, "Randomforest")
bpr=bjpr(testy, pred_test2, "Decision tree")
}
