% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bjroc.R
\name{bjroc}
\alias{bjroc}
\title{Create a ROC with useful information of models in a neat way}
\usage{
bjroc(testy, list1, list2)
}
\arguments{
\item{testy}{Observed outcome values of test dataset}

\item{list1}{list of predicted values from binary classifiers}

\item{list2}{list of model names of binary classifiers}
}
\value{
ROC plot
}
\description{
Create a ROC with useful information of models in a neat way
}
\examples{
testy<-sample(c(0,1), 154, replace=TRUE)
pred_test1<-runif (154, min=0, max=1)
pred_test2<- runif (154, min=0, max=1)
predlist<-list(pred_test1, pred_test2)
modellist<-list("Randomforest","Decision tree")
newroc=bjroc(testy, predlist,modellist)
}
