% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bjshiny.R
\name{bjshiny}
\alias{bjshiny}
\title{Create a Shiny gadget that creates a ROC or PRC according to a user's choice.}
\usage{
bjshiny(testy, x1, x2)
}
\arguments{
\item{testy}{Observed outcome values of test dataset}

\item{x1}{predicted values of a model}

\item{x2}{list of model names of binary classifiers}
}
\value{
Shiny gadget
}
\description{
Create a Shiny gadget that creates a ROC or PRC according to a user's choice.
}
\examples{
testy<-sample(c(0,1), 154, replace=TRUE)
pred_test1<-runif (154, min=0, max=1)
modellist=list("Randomforest","Decision tree")
newshiny=bjshiny (testy, pred_test1,modellist)
}
