% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelmeasure.R
\name{\%>\%}
\alias{\%>\%}
\alias{modelmeasure}
\title{Calculate and collect important performance measures of binary classifiers in a table}
\usage{
modelmeasure(testy, list1, list2)
}
\arguments{
\item{testy}{Observed outcome values of test dataset}

\item{list1}{list of predicted values from binary classifiers}

\item{list2}{list of model names of binary classifiers}
}
\value{
A table that includes measures of model performance
}
\description{
Calculate and collect important performance measures of binary classifiers in a table
}
\examples{
testy<-sample(c(0,1), 154, replace=TRUE)
pred_test1<-runif (154, min=0, max=1)
pred_test2<- runif (154, min=0, max=1)
predlist<-list(pred_test1, pred_test2)
modellist<-list("Randomforest","Decision tree")
newtab=modelmeasure(testy, predlist,modellist)
}
